Move pipeline config
